src/requirements.txt
yfinance
pandas
SQLAlchemy
python-dotenv
